Run as a windows service

+++ main.go
@@ -2,6 +2,7 @@ package main
 
 import (
 	"flag"
+	"fmt"
 	"log"
 	"os"
 	"path/filepath"
@@ -10,7 +11,9 @@ import (
 
 	_ "github.com/mattn/go-sqlite3"
 	"github.com/sunshineplan/metadata"
+	"github.com/sunshineplan/utils"
 	"github.com/vharitonsky/iniflags"
+	"golang.org/x/sys/windows/svc"
 )
 
 // OS is the running program's operating system
@@ -33,8 +36,18 @@ func init() {
 		log.Fatalln("Failed to get self path:", err)
 	}
 	os.MkdirAll(joinPath(dir(self), "instance"), 0755)
-	sqlite = joinPath(dir(self), "instance", "mystocks.db")
+	sqlite = joinPath(dir(self), "instance/mystocks.db")
 	sqlitePy = joinPath(dir(self), "scripts/sqlite.py")
+}
+
+func usage(errmsg string) {
+	fmt.Fprintf(os.Stderr,
+		`%s
+
+usage: %s <command>
+       where <command> is one of install, remove, start, stop.
+`, errmsg, os.Args[0])
+	os.Exit(2)
 }
 
 func main() {
@@ -50,32 +63,58 @@ func main() {
 	iniflags.SetAllowMissingConfigFile(true)
 	iniflags.Parse()
 
+	isIntSess, err := svc.IsAnInteractiveSession()
+	if err != nil {
+		log.Fatalln("failed to determine if we are running in an interactive session:", err)
+	}
+	if !isIntSess {
+		runService(svcName, false)
+		return
+	}
+
 	switch flag.NArg() {
 	case 0:
 		run()
 	case 1:
 		switch flag.Arg(0) {
-		case "run":
+		case "run", "debug":
 			run()
+		case "install":
+			err = installService(svcName, svcDescription)
+		case "remove":
+			err = removeService(svcName)
+		case "start":
+			err = startService(svcName)
+		case "stop":
+			err = controlService(svcName, svc.Stop, svc.Stopped)
 		case "backup":
 			backup()
 		case "init":
-			restore("")
+			if utils.Confirm("Do you want to initialize database?", 3) {
+				restore("")
+			}
 		default:
-			log.Fatalf("Unknown argument: %s", flag.Arg(0))
+			usage(fmt.Sprintf("Unknown argument: %s", flag.Arg(0)))
 		}
 	case 2:
 		switch flag.Arg(0) {
 		case "add":
 			addUser(flag.Arg(1))
 		case "delete":
-			deleteUser(flag.Arg(1))
+			if utils.Confirm(fmt.Sprintf("Do you want to delete user %s?", flag.Arg(1)), 3) {
+				deleteUser(flag.Arg(1))
+			}
 		case "restore":
-			restore(flag.Arg(1))
+			if utils.Confirm("Do you want to restore database?", 3) {
+				restore(flag.Arg(1))
+			}
 		default:
 			log.Fatalf("Unknown arguments: %s", strings.Join(flag.Args(), " "))
 		}
 	default:
-		log.Fatalf("Unknown arguments: %s", strings.Join(flag.Args(), " "))
+		usage(fmt.Sprintf("Unknown arguments: %s", strings.Join(flag.Args(), " ")))
+	}
+	if err != nil {
+		log.Fatalf("failed to %s %s: %v", flag.Arg(0), svcName, err)
 	}
 }
